Moved android library definition logic into AndroidLibrary.kt
