+++ buildSrc/src/main/java/com/stepango/forma/feature/AndroidLibrary.kt
@@ -1,2 +1,43 @@
 package com.stepango.forma.feature
 
+import BuildConfiguration
+import FeatureDefinition
+import applyFrom
+import com.android.build.gradle.LibraryExtension
+import com.stepango.forma.EmptyValidator
+import com.stepango.forma.Library
+import com.stepango.forma.Validator
+import com.stepango.forma.validator
+
+data class AndroidLibraryFeatureConfiguration(
+    val packageName: String,
+    val buildConfiguration: BuildConfiguration,
+    val testInstrumentationRunnerClass: String,
+    val consumerMinificationFiles: Set<String>,
+    val manifestPlaceholders: Map<String, Any> = emptyMap(),
+    val dependencyValidator: Validator = EmptyValidator,
+    val selfValidator: Validator = validator(Library)
+)
+
+fun androidLibraryFeatureDefinition(
+    featureConfiguration: AndroidLibraryFeatureConfiguration
+) = FeatureDefinition(
+    pluginName = "com.android.library",
+    pluginExtension = LibraryExtension::class,
+    featureConfiguration = featureConfiguration,
+    configuration = { extension, feature, _, formaConfiguration ->
+        with(extension) {
+            compileSdkVersion(formaConfiguration.compileSdk)
+
+            defaultConfig.applyFrom(
+                formaConfiguration,
+                feature.testInstrumentationRunnerClass,
+                feature.consumerMinificationFiles,
+                feature.manifestPlaceholders
+            )
+
+            buildTypes.applyFrom(feature.buildConfiguration)
+            compileOptions.applyFrom(formaConfiguration)
+        }
+    }
+)
